requirement.txt
cached-property==1.5.2
Django==2.2
django-redis==4.12.1
django-rest-framework==0.1.0
django-url-filter==0.3.15
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
enum-compat==0.0.3
mysqlclient==2.0.2
PyJWT==1.7.1
pytz==2020.5
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
uWSGI==2.0.19.1
